library/requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
pyparsing==3.0.9
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
